Fix rest of nightly builds failing
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@4477 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ heart/test/TestHeterogeneousConductivities.hpp
@@ -58,9 +58,9 @@ public:
        // HeartConfig::Instance()->SetOdeTimeStep(0.01);
    
    
-        double num_elem_x=8;
-        double num_elem_y=8;
-        double num_elem_z=8;
+        unsigned num_elem_x = 8;
+        unsigned num_elem_y=8;
+        unsigned num_elem_z=8;
         double width=0.1;
         double height=0.1;
         double depth=0.1;
